+++ test/i18next.translate.spec.js
@@ -1,4 +1,4 @@
-// i18next, v1.6.3
+// i18next, v1.6.4_pre
 // Copyright (c)2013 Jan Mühlemann (jamuhl).
 // Distributed under MIT license
 // http://i18next.com
@@ -13,7 +13,7 @@ describe('i18next.translate', function() {
   beforeEach(function() {
     opts = {
       lng: 'en-US',
-      fallbackLng: 'dev', 
+      fallbackLng: 'dev',
       fallbackNS: [],
       fallbackToDefaultNS: false,
       load: 'all',
@@ -273,7 +273,13 @@ describe('i18next.translate', function() {
           en: { translation: {  } },            
           'en-US': { 
             translation: {                      
-              test: { res: 'added __replace__' }
+              test: { res: 'added __replace__',
+                      id: '0',
+                      template: '4',
+                      title: 'About...',
+                      text: 'Site description',
+                      media: ['test'] 
+              }
             } 
           }
         };
@@ -286,9 +292,16 @@ describe('i18next.translate', function() {
         });
   
         it('it should return objectTree', function() {
-          expect(i18n.t('test', { returnObjectTrees: true, replace: 'two' })).to.eql({ 'res': 'added two' });
-          expect(i18n.t('test', { returnObjectTrees: true, replace: 'three' })).to.eql({ 'res': 'added three' });
-          expect(i18n.t('test', { returnObjectTrees: true, replace: 'four' })).to.eql({ 'res': 'added four' });
+          expect(i18n.t('test', { returnObjectTrees: true, replace: 'two' })).to.eql({ 
+            res: 'added two',
+            id: '0',
+            template: '4',
+            title: 'About...',
+            text: 'Site description',
+            media: ['test']
+          });
+          //expect(i18n.t('test', { returnObjectTrees: true, replace: 'three' })).to.eql({ 'res': 'added three' });
+          //expect(i18n.t('test', { returnObjectTrees: true, replace: 'four' })).to.eql({ 'res': 'added four' });
         });
   
       });